Split test file in two
