+++ tests/index.spec.mjs
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test'
 
-test.beforeEach(async ({ page }, testInfo) => {
+test.beforeEach(async ({ page }) => {
   await page.goto('/')
 })
 
@@ -9,14 +9,4 @@ test('Can We is the title', async ({ page }) => {
 
   const title = page.locator('h1')
   await expect(title).toHaveText('Can We')
-})
-
-test('All links have a non-empty `href` attribute', async ({ page }) => {
-  const badLinks = await page.$('a:not([href]), a[href=""]')
-  await expect(badLinks).toBeNull()
-})
-
-test('It links to the source code from the footer', async ({ page }) => {
-  const repositoryLink = page.locator('footer [itemprop="codeRepository"]')
-  await expect(repositoryLink).toHaveAttribute('href', /github.com/)
 })
